Moved feature precomputation integration tests to new file; moved feature precomputation tests to pytest
